Move class from remote to local
This allows us to use a 16-bit version again, but has the downside
that we can no longer check the size class of an arbitrary allocation.
This will break APIs like realloc and malloc_usable_size, since
the local slab descriptors can be cached and not necessarily up-to-date
in memory.

+++ cxlalloc/src/atomic.rs
@@ -78,9 +78,9 @@ macro_rules! impl_convert64 {
 impl_convert64!(u8, u16, u32, u64);
 
 #[repr(C)]
-#[ribbit::pack(size = 8, debug, eq, hash)]
+#[ribbit::pack(size = 16, debug, eq, hash)]
 #[derive(Default)]
-pub struct Version(u8);
+pub struct Version(u16);
 
 impl Version {
     pub fn next(&self) -> Self {
